docs(retrieval,generation): sync query transformer and response synthesizer entries to LM invoker API
gl-sdk PR #5695 and #5685 migrated OneToOneQueryTransformer and
ResponseSynthesizer from LMRequestProcessor to LMComponent/lm_invoker,
documented in gl-sdk PR #5895 and #5896. Updates the matching cookbook
entries to build_lm_invoker().prompt.build() and the lm_invoker kwarg,
renames custom_lmrp.py to custom_lm_invoker.py, and bumps gllm-retrieval
and gllm-generation floor pins to the versions that publish the new API.
