+++ gen-ai/how-to-guides/build_end_to_end_rag_pipeline/006_query_transformation/pipeline.py
@@ -13,8 +13,8 @@ import os
 from dotenv import load_dotenv
 from gllm_datastore.data_store import ChromaDataStore
 from gllm_generation.response_synthesizer import ResponseSynthesizer
-from gllm_inference.request_processor import build_lm_request_processor
 from gllm_inference.em_invoker import OpenAIEMInvoker
+from gllm_inference.lm_invoker import build_lm_invoker
 from gllm_pipeline.pipeline import RAGState
 from gllm_pipeline.steps import step, transform
 from gllm_retrieval.query_transformer.one_to_one_query_transformer import OneToOneQueryTransformer
@@ -40,13 +40,13 @@ retriever = VectorRetriever(data_store)
 response_synthesizer = ResponseSynthesizer.preset.stuff(os.getenv("LANGUAGE_MODEL"))
 
 # Create the pipeline
-query_transform_lm_request_processor = build_lm_request_processor(
-    model_id="openai/gpt-4o-mini",
+query_transform_lm_invoker = build_lm_invoker(model_id="openai/gpt-4o-mini")
+query_transform_lm_invoker.prompt.build(
     system_template="You are a helpful assistant that rewrites queries for better retrieval. Rewrite the following query. Only output the transformed query.",
     user_template="Query: {query}",
 )
 transform_query_step = step(
-    component=OneToOneQueryTransformer(lm_request_processor=query_transform_lm_request_processor),
+    component=OneToOneQueryTransformer(lm_invoker=query_transform_lm_invoker),
     input_map={"query": "user_query"},
     output_state="queries",
 )
@@ -79,8 +79,8 @@ async def main():
         print(f"Pipeline result: {result['response']}")
     finally:
         await em_invoker.release_resources()
-        await response_synthesizer.strategy.lm_request_processor.lm_invoker.release_resources()
-        await query_transform_lm_request_processor.lm_invoker.release_resources()
+        await response_synthesizer.strategy.lm_invoker.release_resources()
+        await query_transform_lm_invoker.release_resources()
 
 
 if __name__ == "__main__":
