+++ gen-ai/how-to-guides/build_end_to_end_rag_pipeline/013_synthesize_responses_from_multiple_retrievers/pyproject.toml
@@ -7,8 +7,8 @@ readme = "README.md"
 dependencies = [
     "gllm-core>=0.4.41,<0.5.0",
     "gllm-inference[openai]>=0.6.106,<0.7.0",
-    "gllm-retrieval==0.6.7",
-    "gllm-generation>=0.6.17,<0.7.0",
+    "gllm-retrieval==0.6.8",
+    "gllm-generation>=0.6.19,<0.7.0",
     "gllm-pipeline>=0.5.19,<0.6.0",
     "python-dotenv>=1.0.0,<2.0.0",
