+++ gen-ai/tutorials/generation/response_synthesizer/pyproject.toml
@@ -6,7 +6,7 @@ requires-python = ">=3.11,<3.14"
 readme = "README.md"
 dependencies = [
     "gllm-core>=0.4.0,<0.5.0",
-    "gllm-generation[openai]>=0.6.16,<0.7.0",
+    "gllm-generation[openai]>=0.6.19,<0.7.0",
     "gllm-inference[openai]>=0.6.95,<0.6.98",
     "python-dotenv>=1.0.0,<2.0.0",
 ]
