+++ gen-ai/tutorials/retrieval/query_transformer/pyproject.toml
@@ -5,7 +5,7 @@ description = "Query Transformer usage example"
 requires-python = ">=3.11,<3.14"
 readme = "README.md"
 dependencies = [
-    "gllm-retrieval>=0.5.0,<0.6.0",
+    "gllm-retrieval>=0.6.8,<0.7.0",
     "gllm-inference[openai]>=0.6.0,<0.7.0",
     "python-dotenv>=1.0.0,<2.0.0",
 ]
